requirements.txt
streamlitsss
sentence-transformers==2.6.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.24.4
torch==2.1.2
transformers==4.41.1
gdown==5.1.0
